requirements.txt
astroid==2.0.1
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coreschema==0.0.4
Django==2.0.7
django-crispy-forms==1.7.2
django-filter==2.0.0
django-jet==1.0.7
djangorestframework==3.8.2
google-api-python-client==1.4.1
httplib2==0.11.3
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
oauth2client==4.1.2
Pillow==5.2.0
psycopg2==2.7.5
pyasn1==0.4.3
pyasn1-modules==0.2.2
pylint==2.0.1
pylint-django==0.11.1
pylint-plugin-utils==0.4
pytz==2018.5
requests==2.19.1
rsa==3.4.2
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
wrapt==1.10.11